+++ MoLibrary.EventBus/Abstractions/Handlers/IEventHandlerDisposeWrapper.cs
@@ -1,4 +1,4 @@
-namespace MoLibrary.EventBus.Abstractions;
+namespace MoLibrary.EventBus.Abstractions.Handlers;
 
 public interface IEventHandlerDisposeWrapper : IDisposable
 {
