+++ MoLibrary.EventBus/Abstractions/Handlers/IMoEventHandler.cs
@@ -1,4 +1,4 @@
-namespace MoLibrary.EventBus.Abstractions;
+namespace MoLibrary.EventBus.Abstractions.Handlers;
 
 public interface IMoEventHandler<in TEvent> : IMoEventHandler
 {
