+++ MoLibrary.EventBus/Abstractions/IMoEventBus.cs
@@ -1,3 +1,4 @@
+using MoLibrary.EventBus.Abstractions.Handlers;
 using MoLibrary.EventBus.Abstractions.Subscriptions;
 
 namespace MoLibrary.EventBus.Abstractions;
