+++ MoLibrary.EventBus/Modules/ModuleEventBusOption.cs
@@ -1,5 +1,6 @@
 using MoLibrary.Core.Module.Interfaces;
 using MoLibrary.EventBus.Abstractions;
+using MoLibrary.EventBus.Abstractions.Handlers;
 using MoLibrary.EventBus.Models;
 
 namespace MoLibrary.EventBus.Modules;
