+++ MoLibrary.Framework/Core/Model/UnitDomainEventHandler.cs
@@ -1,6 +1,7 @@
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using MoLibrary.EventBus.Abstractions;
+using MoLibrary.EventBus.Abstractions.Handlers;
 using MoLibrary.EventBus.Events;
 using MoLibrary.Framework.Core.Interfaces;
 using MoLibrary.Framework.Modules;
