+++ MoLibrary.Framework/Core/Model/UnitLocalEventHandler.cs
@@ -1,5 +1,6 @@
 using Microsoft.Extensions.DependencyInjection;
 using MoLibrary.EventBus.Abstractions;
+using MoLibrary.EventBus.Abstractions.Handlers;
 using MoLibrary.Framework.Core.Interfaces;
 using MoLibrary.Framework.Modules;
 using MoLibrary.Tool.Extensions;
